+++ examples/appdir/src/app/layout.tsx
@@ -1,5 +1,4 @@
 import "./globals.css";
-import "@uploadthing/react/styles.css";
 
 export default function RootLayout({
   children,
