+++ examples/appdir/tailwind.config.js
@@ -4,6 +4,7 @@ module.exports = {
     "./src/pages/**/*.{js,ts,jsx,tsx}",
     "./src/components/**/*.{js,ts,jsx,tsx}",
     "./src/app/**/*.{js,ts,jsx,tsx}",
+    "./node_modules/@uploadthing/react/src/components/**",
   ],
   theme: {
     extend: {
