+++ packages/react/src/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./upload-button";
+export * from "./upload-dropzone";
+export * from "./uploader";
