+++ packages/react/tailwind.config.js
@@ -3,7 +3,6 @@ const colors = require("tailwindcss/colors");
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   content: ["./src/**/*.{js,ts,jsx,tsx,mdx}"],
-  prefix: "ut-",
   theme: {
     fontFamily: {
       display: ["Inter"],
